Move mock_player to the separate header file
